Update compilation to use CMake to build. Only static lib is done.

+++ include/wx4qt/wx/string.h
@@ -2,10 +2,10 @@
 // Name:        wx/buffer.h
 // Purpose:     String implementation
 // Library:     Copied from wxWidgets, modifyed to be used with QT library
-// Author:      StÈphane Ch‚teau
+// Author:      St√©phane Ch√¢teau
 // Modified by:
 // Created:     14/06/2012
-// Copyright:   (c) StÈphane Ch‚teau
+// Copyright:   (c) St√©phane Ch√¢teau
 // Licence:     wxWindows licence
 /////////////////////////////////////////////////////////////////////////////
 #ifndef WX_STRING_HEADER
@@ -54,20 +54,20 @@ public:
     int Len() const                                                                     { return length(); }
     void ToLong(long *_pLong) const                                                     { *_pLong = toLong(); }
     bool IsEmpty() const                                                                { return QString::isEmpty(); }
-    bool IsSameAs(const wxQString &_rstrString,bool bIsCaseSensitive = true) const      { return 0 == QString::compare(_rstrString,bIsCaseSensitive ? Qt::CaseSensitive : Qt::CaseInsensitive); }
+    bool IsSameAs(const wxQString &_rstrString,bool bIsCaseSensitive = true) const      { return 0 == QString::compare(_rstrString, bIsCaseSensitive ? Qt::CaseSensitive : Qt::CaseInsensitive); }
     wxCharBuffer ToUTF8() const                                                         { return toUtf8(); }
-    static wxString FromUTF8(const char *pszText,size_t _stLength)                      { return QString::fromUtf8(pszText,_stLength); }
+    static wxString FromUTF8(const char *pszText,size_t _stLength)                      { return QString::fromUtf8(pszText, static_cast<int>(_stLength)); }
     static wxQString FromUTF8(const char *_pszString)                                   { return QString::fromUtf8(_pszString); }
     int Find(wxChar ch) const                                                           { return QString::indexOf(ch); }
-    wxString Left(size_t count) const                                                   { return QString::left(count); }
-    wxString Right(size_t count) const                                                  { return QString::right(count); }
+    wxString Left(size_t count) const                                                   { return QString::left(static_cast<int>(count)); }
+    wxString Right(size_t count) const                                                  { return QString::right(static_cast<int>(count)); }
     const wxChar *wx_str() const;
     const wxChar *c_str() const;
     const wxChar *utf8_str() const;
-    wxString SubString(size_t nStart = 0,size_t nLen = -1) const                        { return wxString(QString::midRef(nStart,nLen).toString()); }
+    wxString SubString(size_t nStart = 0, size_t nLen = -1) const                       { return wxString(QString::midRef(static_cast<int>(nStart), static_cast<int>(nLen)).toString()); }
     void Replace(const wxChar *_pstrBefore,const wxChar *_pstrAfter)                    { replace(QString(_pstrBefore), QString(_pstrAfter)); }
     // Adds count copies of chPad to the beginning, or to the end of the string (the default).
-    wxString & Pad (size_t count, wxChar chPad= ' ',bool fromRight=true)                { operator=(fromRight ? leftJustified(count, chPad) : rightJustified(count, chPad)); return *this; }
+    wxString & Pad (size_t count, wxChar chPad= ' ', bool fromRight=true)               { operator=(fromRight ? leftJustified(static_cast<int>(count), chPad) : rightJustified(static_cast<int>(count), chPad)); return *this; }
 
     // find first/last occurence of any character not in the set
     size_t find_first_not_of(const wxChar* sz, size_t nStart = 0) const;
